Process change: descriptive git commits
I'm going to start changing the way I use git - now for each commit,
I'll also add a detailed description of the changes and, more
importantly, my thought process of its development. This way the
commit log will hopefully become a developer journal as I develop the
app, and it will be easier to see threads of intent as it grows.

This means I won't be focusing on keeping commits small and incremental.
While there will certainly be small one-off changes, the bulk of
meaningful work will be done in medium-sized chunks that I can address
and explain in a commit message. Instead of worrying about commit size,
I'll try to keep commits focused and complete, meaning I should avoid
splitting a task over multiple commits.

This has the benefit of avoiding the "merges disrupt the story of the
code" problem. It shouldn't matter exactly what order commits are read
in, as long as they're somewhat chronological, since one commit will
describe a whole change unrelated to the changes around it. So it won't
matter that merging a branch interleaves all my commit messages.