Add `npm run start:dev` script, watch file changes
`npm run build` takes a second or two to run, which makes it pretty
annoying to develop when I have to constantly restart the server (and
thus rebuild), so I added a `start:dev` script which simply rebuilds the
CSS, JS, or restarts the server (via nodemon) when those files change.
This way it only rebuilds what it needs to, and is much faster. There's
still a bit of a lag, which I may try to optimize further.

I'd really like to get it to live reload the browser also - that will
probably be my next enhancement.