Add layout and components for home page
The home page consists of 3 parts:

- Page title
- List of posts
- Button for new post

I think this is generally how my page layouts will be - title/nav at the
top, area for content, and then primary actions at the bottom.

I'm starting to really like using the flexbox model to define layouts,
especially in conjuction with RSCSS. I simply give the PostList a flex
of 1, and it magically expands and pushes the title and new post button
to the edges.