Create home page with list of user's posts
For this addition I had to add a couple of methods to UserController to
ensure the login flow was correct. Since I'm using the middleware stack
as a sort of data pipeline, PostController#loadByUser requires
req.data.user to be populated before it's called. In my tests I only
loaded the user from the URL, so I needed to add a similar method that
loaded the logged in user. It may seem a verbose way of structuring
things, but I find it helps me keep track of the data flow - the
pipeline is defined in the route file, which calls controller methods to
load only the data needed for the request.