Add New Post form and View Post pages.
As part of this, I added a util file with some helper methods for
rendering, since that react mess was so awful to paste everywhere. I ran
into the problem of how to render html or send json depending on the
type of request - I settled this by adding `res.renderReact` to wrap a
lot of the weird rendering so that my middleware functions could use it.