Add animation to PostRow component.
This is really cool! Now that I finally figured out how to properly
initialize React on the client-side, I can start adding dynamic
functionality. The first thing I wanted to add was an iOS-like slide
animation to the list elements of the post list, to reveal some context
buttons. Given my previous experience with custom drag-and-drop
javascript, it wasn't too hard to figure out.  I'm especially glad I can
use CSS for all of the actual transitions, which I believe is
resource-efficient.

I will say that now more than ever I'm feeling the benefits of
component-oriented development, both with React and RSCSS. Everything
seems well-contained and predictable so far (with the exception of some
strange hoops I had to go through to get the server's data on the
client) which lets me focus on the interesting development rather than
the architecture.