Add post row buttons and refine post row animation
I've added some buttons when a post row is slid out. The tricky part
with this was defining their position in css, since I wanted to use flex
box for the general layout, and have the buttons slide in from off
canvas on the right. It was a little tricky, but I finally figured out
that I could keep the buttons' width out of the flex calculation by
making them absolutely positioned. After a lot of fiddling with the css,
I got a solution I'm happy with - no hacks, flexible, seamlessly
functional. This whole flexbox/animated sliding adventure has really
cemented my respect for the recent developments in HTML and CSS.